requirements.txt
Flask == 1.0.2
Pillow == 7.2.0
loguru == 0.5.3
matplotlib == 3.1.1
numpy == 1.19.5
pandas == 1.1.1
scikit-learn == 0.24.2
tensorflow == 2.4.1
tqdm == 4.64.0
transformers == 4.18.0
albumentations==1.3.1
ipykernel==6.27.1
ipython==8.15.0
ipywidgets==8.1.1
jupyter==1.0.0
matplotlib-inline==0.1.6
numpy==1.26.0
pandas==2.1.4
Pillow==9.4.0
timm==0.9.12
